Move default_types.rs to tools/ripgrep module
