+++ crates/maple_cli/src/tools/ripgrep/default_types.rs
